requirements.txt
numpy==1.24.2
Pillow==9.5.0
pytest==7.2.2
